Move dataset.ts to the top level.
