+++ src/primitives/circles.ts
@@ -1,7 +1,7 @@
 import { Buffer } from "regl";
 import { CandyGraph } from "../candygraph";
 import { Primitive, NumberArray, NamedDrawCommands } from "../common";
-import { createDataset, Dataset } from "./dataset";
+import { createDataset, Dataset } from "../dataset";
 
 export interface CirclesOptions {
   /** The interior color of the circles. If this value is a single Vector4, it will apply to all the circles. Default [0, 0, 0, 0.5]. */
